+++ packages/tests/src/e2e/sourceCodeContext/sourceCodeContext.spec.ts
@@ -4,6 +4,7 @@
 
 /* eslint-env browser */
 /* global globalThis */
+import type { BundlerName } from '@dd/core/types';
 import { verifyProjectBuild } from '@dd/tests/_playwright/helpers/buildProject';
 import type { TestOptions } from '@dd/tests/_playwright/testParams';
 import { test } from '@dd/tests/_playwright/testParams';
@@ -16,6 +17,7 @@ const { expect, beforeAll, describe } = test;
 
 const SERVICE_NAME = 'test-micro-frontend';
 const SERVICE_VERSION = '1.2.3';
+const UUID_RX = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
 
 const userFlow = async (url: string, page: Page, bundler: TestOptions['bundler']) => {
     // Navigate to our page.
@@ -30,27 +32,35 @@ const getRUMEvents = async (page: Page) => {
     return page.evaluate(() => (globalThis as any).rum_events);
 };
 
+const getDebugIds = async (page: Page): Promise<string[]> => {
+    return page.evaluate(() =>
+        Object.values((globalThis as any)['DD_SOURCE_CODE_CONTEXT'] || {})
+            .map((ctx: unknown) => (ctx as Record<string, unknown>)?.ddDebugId)
+            .filter((id: unknown): id is string => typeof id === 'string'),
+    );
+};
+
+async function build(publicDir: string, suiteName: string, bundlers: BundlerName[]) {
+    const source = path.resolve(__dirname, 'project');
+    const destination = path.resolve(publicDir, suiteName);
+    await verifyProjectBuild(source, destination, bundlers, pluginConfig, { splitting: true });
+}
+
+const pluginConfig = {
+    ...defaultConfig,
+    rum: {
+        sourceCodeContext: {
+            service: SERVICE_NAME,
+            version: SERVICE_VERSION,
+            debugId: true,
+        },
+    },
+};
+
 describe('Source Code Context', () => {
     // Build our fixture project.
     beforeAll(async ({ publicDir, bundlers, suiteName }) => {
-        const source = path.resolve(__dirname, 'project');
-        const destination = path.resolve(publicDir, suiteName);
-        await verifyProjectBuild(
-            source,
-            destination,
-            bundlers,
-            {
-                ...defaultConfig,
-                rum: {
-                    enable: true,
-                    sourceCodeContext: {
-                        service: SERVICE_NAME,
-                        version: SERVICE_VERSION,
-                    },
-                },
-            },
-            { splitting: true },
-        );
+        await build(publicDir, suiteName, bundlers);
     });
 
     test('Should inject DD_SOURCE_CODE_CONTEXT global variable', async ({
@@ -84,13 +94,7 @@ describe('Source Code Context', () => {
         expect(errors).toEqual([]);
     });
 
-    test('Should not throw errors', async ({
-        page,
-        bundler,
-        browserName,
-        suiteName,
-        devServerUrl,
-    }) => {
+    test('Should not throw errors', async ({ page, bundler, suiteName, devServerUrl }) => {
         const errors: string[] = [];
         const testBaseUrl = `${devServerUrl}/${suiteName}`;
 
@@ -147,5 +151,56 @@ describe('Source Code Context', () => {
 
         expect(entryError).toMatchObject({ version: SERVICE_VERSION, service: SERVICE_NAME });
         expect(chunkError).toMatchObject({ version: SERVICE_VERSION, service: SERVICE_NAME });
+    });
+
+    test('Should register a distinct debug_id for a dynamically loaded chunk', async ({
+        page,
+        bundler,
+        suiteName,
+        devServerUrl,
+    }) => {
+        await userFlow(`${devServerUrl}/${suiteName}`, page, bundler);
+
+        const before = await getDebugIds(page);
+
+        // Loading a separate chunk evaluates its own injected snippet.
+        await page.click('#load_chunk');
+        await page.waitForFunction(() => window.chunkLoaded === true);
+
+        const after = await getDebugIds(page);
+
+        // The chunk contributed at least one new, distinct debug_id (per emitted file).
+        expect(after.length).toBeGreaterThan(before.length);
+        const newDebugIds = after.filter((debugId) => !before.includes(debugId));
+        expect(newDebugIds.length).toBeGreaterThanOrEqual(1);
+        for (const debugId of after) {
+            expect(debugId).toMatch(UUID_RX);
+        }
+    });
+
+    test('Should generate the same debug_id across two builds', async ({
+        page,
+        bundler,
+        suiteName,
+        devServerUrl,
+        publicDir,
+        bundlers,
+    }) => {
+        // rspack chunk.contentHash.javascript is non-deterministic across builds when devtool
+        // is enabled, causing different debug IDs each time.
+        test.skip(
+            bundler === 'rspack',
+            'rspack content hash is not deterministic across build directories when devtool is enabled',
+        );
+
+        await build(publicDir, `${suiteName}-rebuild`, bundlers);
+
+        await userFlow(`${devServerUrl}/${suiteName}`, page, bundler);
+        const firstBuildIds = await getDebugIds(page);
+
+        await userFlow(`${devServerUrl}/${suiteName}-rebuild`, page, bundler);
+        const secondBuildIds = await getDebugIds(page);
+
+        expect(firstBuildIds.sort()).toEqual(secondBuildIds.sort());
     });
 });
